refactor unittests, extend & test baseReader, integrate new Cal, bump version
