+++ shepherd_core/setup.cfg
@@ -42,6 +42,7 @@ install_requires =
     pyYAML
     chromalog
     pydantic[email]<2.0.0
+    tqdm
 
 [options.extras_require]
 dev =
