+++ shepherd_core/shepherd_core/__init__.py
@@ -5,22 +5,28 @@ Provides classes for storing and retrieving sampled IV data to/from
 HDF5 files.
 
 """
-from .calibration import raw_to_si
-from .calibration import si_to_raw
+from .data_models.base.calibration import CalibrationCape
+from .data_models.base.calibration import CalibrationEmulator
+from .data_models.base.calibration import CalibrationHarvester
+from .data_models.base.calibration import CalibrationPair
+from .data_models.base.calibration import CalibrationSeries
 from .logger import get_verbose_level
 from .logger import logger
 from .logger import set_verbose_level
 from .reader import BaseReader
 from .writer import BaseWriter
 
-__version__ = "2023.5.5"
+__version__ = "2023.5.6"
 
 __all__ = [
     "BaseReader",
     "BaseWriter",
-    "raw_to_si",
-    "si_to_raw",
     "get_verbose_level",
     "set_verbose_level",
     "logger",
+    "CalibrationCape",
+    "CalibrationSeries",
+    "CalibrationEmulator",
+    "CalibrationHarvester",
+    "CalibrationPair",
 ]
