+++ shepherd_core/tests/data_models/conftest.py
@@ -1,0 +1,8 @@
+import yaml
+from pathlib import Path
+
+
+def load_yaml(file: str):
+    yaml_path = Path(__file__).resolve().with_name(file)
+    with open(yaml_path) as _data:
+        return yaml.safe_load(_data)
