+++ shepherd_core/tests/data_models/test_task_generation.py
@@ -11,14 +11,14 @@ from shepherd_core.data_models import VirtualSource
 from shepherd_core.data_models.task import TestbedTasks
 
 
-def test_task_generation_file():
+def test_task_generation_file(tmp_path: Path):
     path = Path(__file__).with_name("example_config_experiment.yaml")
     xp1 = Experiment.from_file(path)
     tb_tasks = TestbedTasks.from_xp(xp1)
-    tb_tasks.to_file(Path("tbt1.yaml"))
+    tb_tasks.to_file(tmp_path / "tbt1.yaml")
 
 
-def test_task_generation_script():
+def test_task_generation_script(tmp_path: Path):
     hrv = VirtualHarvester(name="mppt_bq_thermoelectric")
 
     target_cfgs = [
@@ -49,4 +49,4 @@ def test_task_generation_script():
     )
 
     tb_tasks = TestbedTasks.from_xp(xperi)
-    tb_tasks.to_file(Path("tbt2.yaml"))
+    tb_tasks.to_file(tmp_path / "tbt2.yaml")
