--- shepherd_core/tests/example_config_emulator.yaml
@@ -1,43 +1,0 @@
-# This is an example configuration for shepherd.
-# Instead of modifying, better make a copy and name it config.yml
-# Behavior: emulate a BQ-Controller for target A, enable io-connection
-mode: emulator
-parameters:
-  input_path: /var/shepherd/recordings/hrv.h5
-  # output_path:
-  # - providing a directory -> file is named emu_timestamp.h5
-  # - for a complete path the filename is not changed except it exists and overwrite is disabled -> emu#num.h5
-  output_path: /var/shepherd/recordings
-  force_overwrite: false
-
-  time_start: 2042-12-13 14:15:16
-  duration: 600
-
-  enable_io: true
-  io_port: A  # either Port A or B
-  pwr_port: A
-
-  voltage_aux: 3.3
-  # aux_voltage options:
-  # - None to disable (0 V),
-  # - 0-4.5 for specific const Voltage,
-  # - "mid" will output intermediate voltage (vsource storage cap),
-  # - true or "main" to mirror main target voltage
-
-  virtual_source: # provide path, name (ie. default, BQ25570) or parameters
-    inherit_from: BQ25570
-    harvester:
-      name: "mppt_bq_solar"
-    V_intermediate_init_mV: 3600
-    V_buck_drop_mV: 0
-    V_output_mV: 3000
-
-  power_tracing:
-    voltage: true
-    current: true
-  gpio_tracing:
-    uart_decode: true
-    uart_baudrate: 9600
-  sys_logging:
-    dmesg: true
-verbose: 2 # serious performance impact for value > 3
