--- shepherd_core/tests/example_config_experiment.yaml
@@ -1,16 +1,0 @@
----
-model: Experiment
-fields:
-  id: 4567
-  name: meaningful Test-Name
-  created: 2023-11-11 11:11:11
-  time_start: 2023-12-12 12:12:12
-  target_configs:
-    - target_IDs: [2001, 2002, 2003, 2004]
-      custom_IDs: [0, 1, 2, 3]
-      energy_env:
-        name: SolarSunny
-      virtual_source:
-        name: diode+capacitor
-      firmware1:
-        name: nrf52_demo_rf
