+++ shepherd_data/shepherd_data/__init__.py
@@ -9,7 +9,7 @@ from shepherd_core import BaseWriter as Writer
 
 from .reader import Reader
 
-__version__ = "2023.5.5"
+__version__ = "2023.5.6"
 
 __all__ = [
     "Reader",
